+++ lib/product_page/product_list_page.dart
@@ -1,31 +1,37 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_realestate_shopping_app/pages/Product_register_page.dart';
-import 'package:flutter_realestate_shopping_app/widgets/Product_card.dart';
+import 'package:flutter_realestate_shopping_app/product_page/product_register_page.dart';
+import 'package:flutter_realestate_shopping_app/product_page/widgets/product_card.dart';
 import '../models/product_model.dart';
 
-class ProductListPage extends StatelessWidget {
-  final List<Product> productList;
-  final Function(Product)? onProductAdded;
+class ProductListPage extends StatefulWidget {
+  const ProductListPage({super.key});
 
-  const ProductListPage({
-    super.key,
-    required this.productList,
-    this.onProductAdded,
-  });
+  @override
+  State<ProductListPage> createState() => _ProductListPageState();
+}
+
+class _ProductListPageState extends State<ProductListPage> {
+  final List<Product> _productList = [];
+
+  void _addProduct(Product product) {
+    setState(() {
+      _productList.add(product);
+    });
+  }
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(title: const Text('매물 목록')),
-      body: productList.isEmpty
+      body: _productList.isEmpty
           ? const Center(
               child: Text('매물이 존재하지 않습니다.', style: TextStyle(fontSize: 16)),
             )
           : ListView.builder(
               padding: const EdgeInsets.all(16),
-              itemCount: productList.length,
+              itemCount: _productList.length,
               itemBuilder: (context, index) {
-                return ProductCard(product: productList[index]);
+                return ProductCard(product: _productList[index]);
               },
             ),
       floatingActionButton: FloatingActionButton(
@@ -34,10 +40,8 @@ class ProductListPage extends StatelessWidget {
             context,
             MaterialPageRoute(builder: (_) => const ProductRegisterPage()),
           );
-          if (newProduct != null &&
-              newProduct is Product &&
-              onProductAdded != null) {
-            onProductAdded!(newProduct); // 새 상품 리스트에 추가
+          if (newProduct != null && newProduct is Product) {
+            _addProduct(newProduct);
           }
         },
         child: const Icon(Icons.add),
